+++ packages/victory-bar/src/victory-bar.js
@@ -64,6 +64,7 @@ class VictoryBar extends React.Component {
         bottomRight: PropTypes.oneOfType([PropTypes.number, PropTypes.func])
       })
     ]),
+    getPath: PropTypes.func,
     horizontal: PropTypes.bool
   };