+++ test/client/spec/victory-core/victory-primitives/point.spec.js
@@ -2,6 +2,7 @@
 
 import React from "react";
 import { shallow } from "enzyme";
+import { assign } from "lodash";
 import Point from "packages/victory-core/src/victory-primitives/point";
 import Path from "packages/victory-core/src/victory-primitives/path";
 import pathHelpers from "packages/victory-core/src/victory-primitives/path-helpers";
@@ -35,7 +36,7 @@ describe("victory-primitives/point", () => {
       "star"
     ].forEach((symbol) => {
       const stub = sandbox.stub(pathHelpers, symbol).returns(`${symbol} symbol`);
-      const props = Object.assign({}, baseProps, { symbol });
+      const props = assign({}, baseProps, { symbol });
       const wrapper = shallow(<Point {...props} />);
       const directions = wrapper.find(Path).prop("d");